feat: create getSummary and refactor App
Created getSummary function to make a
GET request and return the summary as
a JSON object. The App component has
been simplified: Uses getSummary to get
and store the summary.Displays Summary if
there is data, or EmptyGoals if there is
not.The CreateGoal component is always
displayed inside a dialog.